Rename func.py to utils.py
Update file docstrings

Remove trailing space at end of AmpliPIStream.py
